Rename utils folder to core because that's where the meat of the application lies
