+++ core/graph.ts
