+++ core/inheritance.ts
