+++ core/patrimony.ts
