+++ core/tests/inheritance.test.ts
